fix(server-client): stop the GeometryDiagnostics copy drifting from the geometry one
`ParseResponse.geometry_diagnostics` was typed against a private copy of
`GeometryDiagnostics` that had missed three rounds of additions to the
canonical type in `@ifc-lite/geometry`: `schemaVersion`, `worstHosts.bbox`,
`worstHosts.triangleCount` and (as of the previous commit) `oversizedRefDrops`.
The Rust server serialises all of them, so a consumer reading one got a type
error and reached for a cast (#3857).

The copy stays a copy. `@ifc-lite/server-client` ships with an empty
`dependencies` map on purpose - it is the thin HTTP client - and
`@ifc-lite/geometry` drags in `@ifc-lite/wasm` (a cargo/wasm-pack build) plus
`@ifc-lite/data`. A type-only import would not stay type-only either:
declaration emit writes the cross-package reference into `dist/index.d.ts`, so
every consumer would need the geometry package installed just to typecheck.

What changes is that the copy is now checked rather than promised:

- `geometry-diagnostics-types.ts` holds the wire shape, synced field for field
  with the canonical type (the `types.ts` re-export keeps the public surface
  identical, following the `symbolic-types.ts` precedent). This also takes
  `types.ts` from 627 to 596 lines; the module-size row ratchets down.
- `geometry-diagnostics-contract.test.ts` asserts exact structural equality
  against `packages/geometry/src/diagnostics.ts` at COMPILE TIME (type-only
  relative import, no dependency edge, no built dist; precedent
  `packages/sdk/src/sandbox.test.ts`). Assignability would not do: an interface
  with extra properties is still assignable, which is the drift being caught.

Verified by mutation, each one failing `pnpm typecheck` on its own: removing a
field from the copy, adding one to the copy, and adding one to the canonical
type.

`totalUnsupportedItems` / `unsupportedItemsByType` are declared here ahead of
the canonical type so that open PR #3691 and this change do not collide over
the wire shape. They sit in a named allowlist in the contract test, and a
second assertion fails the moment `@ifc-lite/geometry` gains those names, which
is the signal to empty the allowlist.

Closes #3857.
