+++ scripts/module-size-allowlist.txt
@@ -333,7 +333,7 @@
   1082 packages/server-client/src/client.ts
    775 packages/server-client/src/data-model-decoder.ts
    466 packages/server-client/src/parquet-tables.ts
-   627 packages/server-client/src/types.ts
+   596 packages/server-client/src/types.ts
    411 packages/source-dalux/src/http-client.ts
    406 packages/source-dalux/src/provider.ts
    432 packages/source-dropbox/src/provider.ts
